+++ CHANGELOG.md
@@ -734,6 +734,8 @@ Scripting and execution
 -   Add `$provideScriptingForNode`
 -   Converts JSON strings to JSON objects
 -   Add `jsonStringsToJsons` to `@promptbook/utils`
+-   Increase `DEFAULT_MAX_EXECUTION_ATTEMPTS` from `3` -> `10`
+-   Add a unique ID to the error, this error needs to be serialised and deserialised.
 
 ## In pre-release